// web/src/api/models/Layout.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LayoutMonitor } from './LayoutMonitor';
export type Layout = {
    id: string;
    name: string;
    emoji?: string;
    aliases: Array<string>;
    monitors: Array<LayoutMonitor>;
};